refactor: split getArtistKnowledge.ts into separate files following Single Responsibility Principle
